backend/requirements.txt
tortoise-orm
fastapi
uvicorn
python-jose
passlib[bcrypt]
aiofiles
aiohttp
jsonpath
apscheduler